Move ShowCFGCommand to magpiebridge package
